code: add SEM_likelihood function
This function computes likelihood for given data and chosen parameters. For now it computes the likelihood exactly as it has been done in the original code. However most likely this should change since there is a mistake in the way the likelihood was computed. Sigma12 matrix is always prepared based on all of the regressors and not only regressors used in the current model. Hence other regressors are used when the likelihood is computed. This is most likely incorrect and if so should be changed.