code: fill in zeros in sigma matrix if regressor is not in the current model
This change finally fixes the issue with the sigma matrix. When we have a regressor which is not included in the current model, then sigma12 matrix should have zeros in columns corresponding to this regressor. Otherwise we are computing likelihood based on values of regressors which should not be in the current model. This is especially important during optimisation - without this change we are able to optimise likelihood by changing parameters corresponding to regressors not included in the model. Therefore we obtain incorrect maximum as well as incorrect values of maximising parameters.