fix:  exclude parameters corresponding to regressors not included in the current model
Up to this point matrix sigma12 contained non-zero parameters for regressors not included in the current model. This generated false results, because during maximisation it was possible to optimise likelihood with respect to these parameters. Now these parameters are excluded and the matrix sigma12 contains 0s in columns corresponding to the currently excluded regressors.