code: prior depends on correlation matrix in the gauss code
Priors are now multiplied by square root of the determinant of the correlation matrix computed for regressors used in the current model. If there are less than 2 regressors the determinant is assumed to be equal to one. Correlations are computed based on the values of regressors for all time periods and all countries.