fix: use only regressors from the current model when creating sigma matrix
So far the aim was to replicate Moral-Benito's results. However original code contained a mistake: Sigma12 matrix contained always the same number of parameters no matter which model was currently considered. This is obvoiusly incorrect, since then we can optimise likelihood using regressors which are outside of the currently considered model. This commit aims to fix that issue.