fix: delete zero rows from sigma matrix and do not use all regressors in Y2 during BMA
It is not enough to just correct Sigma12 matrix, because Y2 matrix is also used to compute likelihood. Therefore we will obtain incorrect values of likelihood. This might lead also to finding incorrect maximising arguments. Hence we have to remove elements of both matrices Sigma12 and Y2 when we are considering different models during Bayesian Model Averaging (BMA). Namely, we have to remove the elements connected to regressors which are not considered in the currently considered model.