fix: remove connection to regressors outside of current model in gauss code
Fixing the issue with Sigma12 matrix requires not only putting zeros in the rows and columns connected with regreressors outside of the current model within the Sigma12 matrix. Other matrices (e.g. B) as well as likelihood computation have to be revised. This commit changes dimensions of matrices used so that only regressors inside the current model are included in the matrices (except B matrix, since it is enough to use zeros for this matrix).