code: allow to use SEM_likelihood directly with dataframe
For many users it might be easier to provide a dataframe to SEM_likelihood without preparing the necessary matrices beforehand. Now this functionality is available.

However, if we call SEM_likelihood repeatedly at the same point (e.g within the call to optim function) it is better to provide matrices directly to the SEM_likelihood so that they are not recalculated at each call to the function. Hence the old way of calling SEM_likelihood is still possible.