code: reset seed at each iteration
To compare resutls with corrected version of likelihood we have to set the seed at each iteration, because optmum function has random behaviour in it. If we set the seed before the entire loop, we end up at different pseudo-random point for each model using different versions of likelihood (because random functions are called different number of times). With seed set at each iteration we will be able to check if the corrected version of likelihood generates the same results for the last model (it should, since we use all regressors in the last model)