code: introduce corrections to likelihood
This corrections fix again the issue with too many regressors considered in the likelihood (precisely sigma12 matrix). This was in depth explained in previous commits. Now the same fix is applied but in such a way that we can compare the results with two previous commits which show the results for the original code.