code: use the same likelihood form for all models
In order to conduct Bayesian Model Averaging (BMA) all likelihoods must have the same form and differ only in the value of beta and phi1 parameter vectors which determine dependence of dependent variable on specific regressors. Namely for regressors outside of the current model we impose that beta and phi1 is equal to 0 at corresponding coordinates, i.e. these regressors have no impact direct impact on dependent variable. However, the rest of the likelihood stays the same and we still have to optimise over parameters corresponding to the error terms of regressors outside of the current model (outside meaning regressors which are assumed to have no direct impact on dependent variable). This has to be done in this manner, because only then we can use likelihood ratio statistic to compare different models.

This is why previous commits which applied a so-called fix to the sigma matrix were in truth introducing a mistake in the overall BMA analysis.