refactor: rename grad parameter to per_entity
This name better describes the idea behind this alternative computation. Namely with per_entity set to TRUE SEM_likelihood returns vector with log-likelihoods computed for each entity from the dataset (e.g. countries). If we sum up these values we again get the total log-likelihood.

The name "grad" was an artifact from Moral-Benito code. Most likely the nomenclatture connected to gradient was used, because this vector log-likelihood is then used to compute Gmat matrix which is in turn computed as gradient of the mentioned vector log-likelihood. Do not know yet why this is so.