fix: subtract constant to normalise likelihood
The models considered in this scenario are not nested. Hence to compare likelihood we have to normalise them. Note that simply normalising the likelihoods has not been proven (at least as far as I know so far) to generate the same statistics as in the nested models case. It is just common sense that such comparison gives more reasonable results. Or, to put it in other words, it makes more sense from the bayesian perspective.

It will be possible to get some more precise corrolaries after simulations are run for such scenarios. Nevertheless normalising the likelihoods seems like a good idea to get some insights how models behave compared to each other.