fix: apply changes to grad likelihood
This commit applies the same changes which were made in few recent commits, but to the gradient of likelihood. Still have to rethink whether this is how this should be applied: where split into elements and where to divide by n. However this does not change the final result (the question whether grad likelihood function is even needed is once again raised here) and so we apply this changes in this manner for now.