feat: create file for running MC simulations
The idea is to compare two methods. One, where only beta parameters, i.e. the ones responsible for linear dependence are set to zero when a regressor is considered outside of the model. Second model, where a removed regressor is completely removed from the model. Running multiple Monte Carlo simulations should show which method is better