feat: added term for trace part simplification
In the Likelihood-based Estimation of Dynamic Panels with Predetermined Regressors Moral-Benito simplifies the formula for likelihood (see AppendixA.1). When transitioning from equation (47) to equation (48), in step 2, the term trace{HG_22} is dropped, because it can be assumed to be constant from Moral-Benito perspective. To get the exact value of the likelihood we have to take this term into account.