feat: simplified data preparation framework
I've removed data_prep and created a simple version of feature standardization. Now the user can simply choose the subset of columns to which the transformation should be applied as well as whether there should be a grouping column. It's also possible to choose if we want to apply both demeaning and rescaling or just demeaning