revised README after adding new features (#89)
* docs: first draft of the new README

* feat: remove safeMakeCluster

In the end I don't like the idea of creating a wrapper function just for the purpose of testing on CRAN. Especially since it is completely unrelated to our package. Maybe there is a way to do it in a reasonable manner, but definitely not like we had it now - as a part of data_preparation.R

* docs: hide long outputs

* docs: split long lines and reformat lists

* feat: simplified data preparation framework

I've removed data_prep and created a simple version of feature standardization. Now the user can simply choose the subset of columns to which the transformation should be applied as well as whether there should be a grouping column. It's also possible to choose if we want to apply both demeaning and rescaling or just demeaning

* docs: set myself as the first author

* docs: change number of best and top models so it makes sense for small data frame