fix: take into account sign when computing log det
arma::log_det returns a finite number even when the determinant is negative, because it computes log|det| plus a sign. The sign was so far discarded. This is hopefully the root cause of the error we were getting, i.e., that sometimes we had parameters in the model but standard deviations were turning up as NAs